+++ x/slashing/test_common.go
@@ -93,8 +93,9 @@ func testAddr(addr string) sdk.Address {
 func newTestMsgCreateValidator(address sdk.Address, pubKey crypto.PubKey, amt sdk.Int) stake.MsgCreateValidator {
 	return stake.MsgCreateValidator{
 		Description:    stake.Description{},
+		DelegatorAddr:  address,
 		ValidatorAddr:  address,
 		PubKey:         pubKey,
-		SelfDelegation: sdk.Coin{"steak", amt},
+		Delegation: sdk.Coin{"steak", amt},
 	}
 }