+++ x/stake/types/msg_test.go
@@ -70,7 +70,7 @@ func TestMsgEditValidator(t *testing.T) {
 }
 
 // test ValidateBasic for MsgSurrogateCreateValidator
-func TestMsgSurrogateCreateValidator(t *testing.T) {
+func TestMsgOnBehalfOfCreateValidator(t *testing.T) {
 	tests := []struct {
 		name, moniker, identity, website, details string
 		surrogateAddr                             sdk.Address
@@ -82,7 +82,7 @@ func TestMsgSurrogateCreateValidator(t *testing.T) {
 		{"basic good", "a", "b", "c", "d", addr1, addr2, pk2, coinPos, true},
 		{"partial description", "", "", "c", "", addr1, addr2, pk2, coinPos, true},
 		{"empty description", "", "", "", "", addr1, addr2, pk2, coinPos, false},
-		{"empty surrogate address", "a", "b", "c", "d", emptyAddr, addr2, pk2, coinPos, false},
+		{"empty delegator address", "a", "b", "c", "d", emptyAddr, addr2, pk2, coinPos, false},
 		{"empty validator address", "a", "b", "c", "d", addr1, emptyAddr, pk2, coinPos, false},
 		{"empty pubkey", "a", "b", "c", "d", addr1, addr2, emptyPubkey, coinPos, true},
 		{"empty bond", "a", "b", "c", "d", addr1, addr2, pk2, coinZero, false},
@@ -92,7 +92,7 @@ func TestMsgSurrogateCreateValidator(t *testing.T) {
 
 	for _, tc := range tests {
 		description := NewDescription(tc.moniker, tc.identity, tc.website, tc.details)
-		msg := NewMsgSurrogateCreateValidator(tc.surrogateAddr, tc.validatorAddr, tc.validatorPubKey, tc.bond, description)
+		msg := NewMsgOnBehalfOfCreateValidator(tc.surrogateAddr, tc.validatorAddr, tc.validatorPubKey, tc.bond, description)
 		if tc.expectPass {
 			require.Nil(t, msg.ValidateBasic(), "test: %v", tc.name)
 		} else {